+++ src/main/resources/db/changelog/changelogs/changeset-002-create-products.sql
@@ -5,7 +5,7 @@
 --preConditions onFail:MARK_RAN
 
 --sql
-CREATE TABLE IF NOT EXISTS public.Products (
+CREATE TABLE IF NOT EXISTS Products (
 id UUID PRIMARY key default uuid_generate_v4() ,
 name VARCHAR(255),
 description TEXT,