+++ src/main/resources/db/changelog/changelogs/changeset-003-create-orders.sql
@@ -5,13 +5,13 @@
 --preConditions onFail:MARK_RAN
 
 --sql
-CREATE TABLE IF NOT EXISTS public.Orders (
+CREATE TABLE IF NOT EXISTS Orders (
 id UUID PRIMARY KEY DEFAULT uuid_generate_v4(),
 user_id UUID,
 product_id UUID,
 quantity INTEGER,
 order_date TIMESTAMP,
 status VARCHAR,
-FOREIGN KEY (user_id) REFERENCES public.users(id),
-FOREIGN KEY (product_id) REFERENCES public.products(id)
+FOREIGN KEY (user_id) REFERENCES users(id),
+FOREIGN KEY (product_id) REFERENCES products(id)
 );